feat: add recent equation display and update logic
Added a display container in HTML for formatting with a new div to hold the most recent equation.
Updated JS to show the most recent equation when operate is ran.
Made a new variable to hold the DOM element recent-equation.
Added logic to the controlHandle() so that they reset the full display. Including the recent equation.
Updated CSS so that the display properly shows the equation.